Add files via upload
This repository uses a pickel dataset named 'all_banks' which is provided here too. A pickle dataset, in layman terms serializes and deserializes the dataset.  This dataset contains information of finance reports of several US banks and we are using that to do some data visualisation  on them and practice.